requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
Pillow>=9.5.0
